Stop webpack from swallowing errors. Closes #91
Webpack dev middleware has a few config options for error reporting and stat
output. The noInfo option is supposed to limit output to errors and warnings,
the quiet option is supposed to supress all output, and the stats option which,
from everything I've seen, can take an 'error-only' value, that when used, does
what you'd expect. The strange thing is that I've seen this in a few blog
posts/issues, but if you look in both webpack-dev-middleware and
webpack-hot-middleware, you don't see anything looking for that value.

At any rate, to test this out, just require a module that doesn't exist
somewhere within /client.